+++ lib/decanter/parser/key_value_splitter_parser.rb
@@ -6,10 +6,12 @@ module Decanter
       ITEM_DELIM = ','
       PAIR_DELIM = ':'
 
-      parser do |_name, val, options|
+      parser do |val, options|
         item_delimiter = options.fetch(:item_delimiter, ITEM_DELIM)
         pair_delimiter = options.fetch(:pair_delimiter, PAIR_DELIM)
-        val.split(item_delimiter).reduce({}) { |memo, pair| memo.merge(Hash[*pair.split(pair_delimiter)]) }
+        val.split(item_delimiter).reduce({}) do |memo, pair|
+          memo.merge(Hash[*pair.split(pair_delimiter)])
+        end
       end
     end
   end