+++ spec/decanter/parser/key_value_splitter_parser_spec.rb
@@ -2,12 +2,19 @@
 
 require 'spec_helper'
 
-describe 'KeyValueSplitterParser' do
-  let(:parser) { Decanter::Parser::KeyValueSplitterParser }
+describe Decanter::Parser::KeyValueSplitterParser do
+  subject { described_class.parse(arg, opts) }
+  let(:opts) { {} }
 
   describe '#parse' do
-    it 'returns an array with the split pairs' do
-      expect(parser.parse('foo', 'foo:bar,baz:foo')).to match('foo' => 'bar', 'baz' => 'foo')
+    let(:arg) { 'foo:bar,baz:quux' }
+    it { is_expected.to match('foo' => 'bar', 'baz' => 'quux') }
+
+    context 'with the delimter option specified' do
+      let(:arg) { 'fooXbarYbazXquux' }
+      let(:opts) { { item_delimiter: 'Y', pair_delimiter: 'X' } }
+
+      it { is_expected.to match('foo' => 'bar', 'baz' => 'quux') }
     end
   end
 end