+++ spec/decanter/parser/pass_parser_spec.rb
@@ -2,12 +2,11 @@
 
 require 'spec_helper'
 
-describe 'PassParser' do
-  let(:parser) { Decanter::Parser::PassParser }
+describe Decanter::Parser::PassParser do
+  subject { described_class.parse(arg) }
+  let(:arg) { '(12)2-21/19.90' }
 
   describe '#parse' do
-    it 'lets anything through' do
-      expect(parser.parse(:foo, '(12)2-21/19.90')).to match(foo: '(12)2-21/19.90')
-    end
+    it { is_expected.to eq arg }
   end
 end