+++ spec/decanter/parser/value_parser_spec.rb
@@ -2,14 +2,14 @@
 
 require 'spec_helper'
 
-describe 'ValueParser' do
-  let(:value_parser) { Class.new(Decanter::Parser::ValueParser) }
+describe Decanter::Parser::ValueParser do
+  let(:value_parser) { Class.new(described_class) }
 
   it 'calls the parser' do
     parser = ->(_a, _b, _c) { { a: 'b' } }
     allow(parser).to receive(:call).and_return({})
-    value_parser.parser &parser
-    value_parser._parse(:first_name, nil)
+    value_parser.parser(&parser)
+    value_parser._parse(nil)
     expect(parser).to have_received(:call).with(nil, {})
   end
 end